reduce usage of blockAddress() function, use Block.Address instead
