+++ pkg/apply_block.go
@@ -1,0 +1,6 @@
+package pkg
+
+type ApplyBlock interface {
+	Block
+	Apply() error
+}
