+++ pkg/custom_decode_base.go
@@ -1,0 +1,7 @@
+package pkg
+
+import "github.com/hashicorp/hcl/v2"
+
+type CustomDecodeBase interface {
+	Decode(*HclBlock, *hcl.EvalContext) error
+}
