+++ pkg/plan_block.go
@@ -1,5 +1,6 @@
 package pkg
 
 type PlanBlock interface {
+	Block
 	ExecuteDuringPlan() error
 }
